[#12] pass only manager to other screens

+++ SIGMA/src/main/kotlin/sigma/businessLogic/impl/managers/StatisticsManager.kt
@@ -2,8 +2,8 @@ package sigma.businessLogic.impl.managers
 
 class StatisticsManager {
     companion object {
-        fun getResult(dayOfMonth: Int): Double {
-            return -1.0
+        fun getResult(dayOfMonth: Int): Float {
+            return -1f
         }
     }
